chore: 미사용 open.piki.day 랜딩 서브도메인 제거
인스타 링크를 `piki.day/open` 으로 직접 걸기로 해서 서브도메인 경로를 걷어낸다.

- `proxy.ts` 의 `open.*` 루트 rewrite 제거
- `isLandingHost` 삭제 — 앱의 associatedDomains 에 등록된 적 없는 호스트였다
- `landingHost.ts` → `serviceHost.ts` 로 이름 변경. 남은 `toServiceHost` 는
  Vercel 프리뷰 등 미등록 호스트를 프로덕션으로 정규화하는 역할만 한다
