+++ apps/web/src/app/open/_components/OpenLanding.tsx
@@ -14,7 +14,7 @@ type OpenLandingProps = {
   landingEnv: LandingEnvT;
   /** 검증된 내부 경로 */
   target: string;
-  /** 실제 서비스 오리진 (piki.day) — 랜딩 오리진과 크로스 도메인이어야 UL 이 산다 */
+  /** 앱 링크가 발동하는 서비스 오리진 (piki.day) — 프리뷰 등 미등록 호스트는 여기서 정규화된다 */
   serviceOrigin: string;
   source: string | null;
 };
@@ -61,7 +61,7 @@ function OpenLanding({ landingEnv, target, serviceOrigin, source }: OpenLandingP
     return () => clearTimeout(timer);
   }, [appOpenUrl, webUrl, platform]);
 
-  /** 함정 1 — JS 이동이라야 크로스 도메인 UL 이 발동하지 않고 웹으로 간다 */
+  /** `<Link>` 가 아닌 JS 이동 — UL 을 건드리지 않고 확실히 웹으로만 보낸다 */
   const handleWebContinueClick = () => {
     logAnalyticsEvent(ANALYTICS_EVENT.LANDING_WEB_CONTINUE, { platform });
     window.location.href = webUrl;
