+++ apps/web/src/proxy.ts
@@ -8,7 +8,6 @@ import { QUERY_ACTION } from './consts/queryAction';
 import { ROUTES } from './consts/route';
 import { getApiErrorStatus } from './utils/apiError';
 import { getRouteType } from './utils/getRouteType';
-import { isLandingHost } from './utils/landingHost';
 import { getLoginPath } from './utils/loginRedirect';
 import { isWebview } from './utils/webBridge';
 
@@ -238,13 +237,6 @@ const handleSessionExpired = (request: NextRequest) => {
 
 export const proxy = async (request: NextRequest) => {
   const { pathname, search } = request.nextUrl;
-
-  /** 랜딩 서브도메인(open.*)의 루트 진입만 랜딩으로 rewrite — 인스타에는 open.piki.day 하나만 건다 */
-  if (pathname === '/' && isLandingHost(request.headers.get('host') ?? '')) {
-    const landingUrl = request.nextUrl.clone();
-    landingUrl.pathname = ROUTES.OPEN;
-    return NextResponse.rewrite(landingUrl);
-  }
 
   const routeType = getRouteType(pathname);
